+++ javascript/src/ErrorList.jsx
@@ -1,5 +1,5 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React from 'react'
+import PropTypes from 'prop-types'
 
 
 const styles = {
@@ -22,4 +22,4 @@ export default function ErrorList(props){
 
 ErrorList.propTypes = {
   message: PropTypes.string.isRequired,
-};
+}