+++ javascript/tests/LoginModal.test.js
@@ -6,7 +6,7 @@ import { mount, shallow } from 'enzyme'
 import Enzyme from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
 
-import { LoginModal } from '../src/LoginModal'
+import LoginModal from '../src/LoginModal'
 import ErrorList from '../src/ErrorList'
 
 import Services from '../src/Services'
@@ -16,8 +16,13 @@ Enzyme.configure({ adapter: new Adapter() })
 
 jest.mock('../src/Services');
 
-const formUrl = 'http://www.example.com'
-const csrfToken = '123'
+const defaultProps = {
+  csrfToken: '123',
+  handleClose: function () {},
+  loginUrl: 'http://www.example.com',
+}
+
+
 const createEvent = () => ({ preventDefault: jest.fn() })
 
 
@@ -29,41 +34,32 @@ afterEach(() => {
   jest.useRealTimers()
 })
 
-test('Modal opens and closes on link click', () => {
-  const component = shallow(<LoginModal loginUrl={formUrl} csrfToken={csrfToken} />)
+test('Modal handles close', () => {
   const event = createEvent()
-
-  // when the user clicks the button
-  act(() => {
-    component.find('#header-sign-in-link').simulate('click', event)
-  })
-
-  // then the modal s open
-  expect(component.find(Modal).prop('isOpen')).toEqual(true)
-  expect(event.preventDefault).toBeCalled()
-
+  const props = {...defaultProps, handleClose: jest.fn()}
+  const component = shallow(<LoginModal {...props} />)
   // when the user clicks the close button
   act(() => {
     component.find(Modal).find(".link").simulate('click', event)
   })
 
-  // then the modal is closed
-  expect(component.find(Modal).prop('isOpen')).toEqual(false)
-
+  // then the handler is closed
+  expect(event.preventDefault).toHaveBeenCalled()
+  expect(props.handleClose).toHaveBeenCalled()
 })
 
 test('Modal shows error message', () => {
   // when there is an error
-  const component = shallow(
-    <LoginModal loginUrl={formUrl} csrfToken={csrfToken} isOpen={true} errorMessage={'some error'} />
-  )
+  const props = {...defaultProps, errorMessage: 'some error', isOpen: true }
+  const component = shallow(<LoginModal {...props} />)
   // then the validation message is displayed
   expect(component.find(ErrorList).prop('message')).toEqual('some error')
 })
 
 test('Modal form elements are disabled while in progress', () => {
   // when the form submission is in progress
-  const component = mount(<LoginModal loginUrl={formUrl} csrfToken={csrfToken} isOpen={true} isInProgress={true} />)
+  const props = {...defaultProps, isOpen: true, isInProgress: true }
+  const component = mount(<LoginModal {...props} />)
   // then the form elements are disabled
   expect(component.find('input[name="username"]').getDOMNode().disabled).toEqual(true)
   expect(component.find('input[name="password"]').getDOMNode().disabled).toEqual(true)
@@ -73,7 +69,8 @@ test('Modal form elements are disabled while in progress', () => {
 
 test('Modal form elements are not disabled while not in progress', () => {
   // when the form submission is in progress
-  const component = mount(<LoginModal loginUrl={formUrl} csrfToken={csrfToken} isOpen={true} isInProgress={false} />)
+  const props = {...defaultProps, isOpen: true, isInProgress: false }
+  const component = mount(<LoginModal {...props} />)
   // then the form elements are disabled
   expect(component.find('input[name="username"]').getDOMNode().disabled).toEqual(false)
   expect(component.find('input[name="password"]').getDOMNode().disabled).toEqual(false)
@@ -98,8 +95,9 @@ describe('Modal end to end', () => {
     // given the credentials are incorrect
     const event = createEvent()
     Services.checkCredentials.mockImplementation(() => Promise.reject('An erorr occured'));
-
-    const component = mount(<LoginModal loginUrl={formUrl} csrfToken={csrfToken} isOpen={true} />)
+ 
+    const props = {...defaultProps, isOpen: true}
+    const component = mount(<LoginModal {...props} />)
 
     act(() => {
       component.find('input[name="username"]').getDOMNode().value = 'username'
@@ -120,8 +118,8 @@ describe('Modal end to end', () => {
     // given the credentials are correct
     const event = createEvent()
     Services.checkCredentials.mockImplementation(() => Promise.resolve());
-
-    const component = mount(<LoginModal loginUrl={formUrl} csrfToken={csrfToken} isOpen={true} />)
+    const props = {...defaultProps, isOpen: true}
+    const component = mount(<LoginModal {...props} />)
 
     act(() => {
       component.find('input[name="username"]').getDOMNode().value = 'username'