+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 ### Implemented enhancements
 - MVP-79 - Add react build tool
 - MVP-52 - Add login and sign up modals
+- MVP-53 - Add login via linked and google